O Schema da Person agora fica em um arquivo próprio
-- O Schema e a interface da Person agora ficam em um módulo 'schema' dentro da person
-- O index da person agora compila o schema e monta os métodos estáticos
