+++ src/db/models/person/index.ts
@@ -1,51 +1,10 @@
-import { ObjectId } from 'mongodb'
-import mongoose, { Document, Schema, Model } from '../../mongoose'
-import currenciesSchema from './currencies'
-import credentialsSchema from './credentials'
-import type { Currencies } from './currencies'
-import type { Credentials } from './credentials'
-
-/**
- * Interface do documento 'person', da collection 'people', que contém
- * informações relacionadas ao usuários
- */
-export interface PersonDoc extends Document {
-	_id: ObjectId
-	/** O email do usuário */
-	email: string
-	/** Sub-documento de credenciais */
-	credentials: Credentials
-	/** Informações de currencies desse usuário */
-	currencies: Currencies
-}
-
-/**
- * Schema do documento de usuários
- */
-const PersonSchema = new Schema({
-	email: {
-		type: String,
-		trim: true,
-		lowercase: true,
-		unique: true,
-		required: true,
-	},
-	credentials: credentialsSchema,
-	currencies: currenciesSchema
-})
-
-PersonSchema.pre('validate', function(this: PersonDoc) {
-	// Ao criar o documento, props de sub-schemas serão undefined
-	if (!this.isNew) return
-	if (typeof this.currencies == 'undefined')
-		// @ts-expect-error Mongoose irá automaticamente preencher o subdocumento
-		this.currencies = {}
-	if (typeof this.credentials == 'undefined')
-		// @ts-expect-error Mongoose irá automaticamente preencher o subdocumento
-		this.credentials = {}
-})
-
+import mongoose, { Model } from '../../mongoose'
+import PersonSchema from './schema'
 import * as balanceOperations from './balancesOps'
+import type { PersonDoc as PersonDocument } from './schema'
+
+// Re-exporta a interface do documento da person
+export type PersonDoc = PersonDocument
 
 /** Interface do Model da Person, com os métodos estáticos do mesmo */
 interface PersonModel extends Model<PersonDoc> {
